🐛 fix(grid): fix the abnormal closing of editing components when the cursor moves
