+++ packages/grid/src/components/index.ts
@@ -1,8 +1,10 @@
+export * from './cell-editors/abstract-cell-editor.component';
 export * from './cell-editors/date/date-editor.component';
 export * from './cell-editors/link/link-editor.component';
 export * from './cell-editors/number/number-editor.component';
 export * from './cell-editors/progress/progress-editor.component';
 export * from './cell-editors/rating/rating-editor.component';
+export * from './cell-editors/select/select-editor.component';
 export * from './cell-editors/text/text-editor.component';
 export * from './cell-views/select/option.component';
 export * from './field-property-editor/field-property-editor.component';
