+++ packages/grid/src/utils/index.ts
@@ -5,6 +5,7 @@ export * from './get-placeholder-cells';
 export * from './get-text-width';
 export * from './image-cache';
 export * from './os';
+export * from './position';
 export * from './style';
 export * from './text-measure';
 export * from './visible-range';
